timelineの処理を別のstateに分割
